requirements.txt
langgraph==1.0.7 
langgraph-prebuilt=1.0.8 
langgraph-checkpoint-postgres==3.0.4 
langgraph-checkpoint-sqlite==3.0.3 
langchain==1.2.10 
langchain-core==1.2.14 
langchain-openai==1.1.10 
langchain-community==0.4.1 
tavily-python==0.5.0 
chromadb==0.5.18 
psycopg[binary]
redis==5.1.1 
pymupdf==1.24.11 
python-docx==1.1.2 
sendgrid==6.11.0 
streamlit==1.39.0 
python-dotenv==1.0.1 
pydantic==2.9.2 
tiktoken==0.8.0